pulsegen now seems to work
lots of changes-- formatting, toplevel testbench works on post-route
now (with appropriate vendorlib hacking).

Still need to get "deadtime" when we've clocked out the last LED.
Also there's a warning that we read/write the same address that
could be fixed by nailing the raddr bus to 0x1ff when we're not using
it.